+++ app.vue
@@ -1,0 +1,13 @@
+<template>
+  <NuxtLayout>
+    <NuxtPage />
+  </NuxtLayout>
+</template>
+
+<script setup lang="ts">
+import {useHead} from "#head";
+
+useHead({
+  titleTemplate: '%s - NuxtStuff',
+})
+</script>