+++ CHANGELOG.md
@@ -6,6 +6,9 @@
   - nodejs: 16.14.0
   - yarn: 1.22.17
 - Add default layout
+- Update Twitter and GitHub icon links
+- Add a TextLink component
+- Complete upgrade from Nuxt 2 to 3
 
 ## 0.1.2